refactor: move some adapters into their respective package
