+++ kitsu/src/main/java/com/chesire/malime/kitsu/api/search/SeriesModelAdapter.kt
@@ -1,8 +1,7 @@
-package com.chesire.malime.kitsu.adapters
+package com.chesire.malime.kitsu.api.search
 
 import com.chesire.malime.core.models.ImageModel
 import com.chesire.malime.core.models.SeriesModel
-import com.chesire.malime.kitsu.api.search.SearchResponse
 import com.squareup.moshi.FromJson
 
 @Suppress("unused")
