+++ kitsu/src/test/java/com/chesire/malime/kitsu/live/LiveKitsuUserTests.kt
@@ -3,7 +3,7 @@ package com.chesire.malime.kitsu.live
 import com.chesire.malime.core.Resource
 import com.chesire.malime.kitsu.adapters.ImageModelAdapter
 import com.chesire.malime.kitsu.adapters.RatingSystemAdapter
-import com.chesire.malime.kitsu.adapters.UserModelAdapter
+import com.chesire.malime.kitsu.api.user.UserModelAdapter
 import com.chesire.malime.kitsu.api.user.KitsuUser
 import com.chesire.malime.kitsu.api.user.KitsuUserService
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
